requirements.txt
langchain
langchain_core
langchain_community
faiss-cpu
langchain_ollama
python-dotenv
SQLAlchemy
python-dotenv
mysql-connector-python
langchain-groq
streamlit
crewai[tools]
crewai
litellm
validators
youtube-transcript-api
pytube
unstructured


